Handle basic command line arguments for the engine
